+++ playback/src/backends/rusty/sink.rs
@@ -6,12 +6,11 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use parking_lot::{Mutex, RwLock};
-use rodio::{queue, source::Done, Source};
+use rodio::{queue, Source};
 use rodio::{OutputStreamHandle, PlayError};
 
-use super::source::SampleType;
-#[allow(unused_imports)] // used for "rusty-soundtouch"
 use super::source::SourceExt as _;
+use super::source::{SampleType, SpecificType};
 use super::PlayerInternalCmd;
 use crate::PlayerCmd;
 
@@ -54,6 +53,18 @@ struct Controls {
     to_clear: Mutex<u32>,
     /// The current position in the currently playing source (may be off by a few milliseconds).
     position: RwLock<Duration>,
+}
+
+/// Options to apply to a specific source
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct SourceOptions {
+    pub soundtouch: bool,
+}
+
+impl Default for SourceOptions {
+    fn default() -> Self {
+        Self { soundtouch: true }
+    }
 }
 
 #[allow(dead_code)]
@@ -101,7 +112,7 @@ impl Sink {
     /// Appends a sound to the queue of sounds to play.
     #[inline]
     #[allow(clippy::cast_possible_wrap)]
-    pub fn append<S>(&self, source: S)
+    pub fn append<S>(&self, source: S, options: &SourceOptions)
     where
         S: Source<Item = SampleType> + Send + 'static,
     {
@@ -118,7 +129,7 @@ impl Sink {
         let progress_tx = self.picmd_tx.clone();
         let source = source
             .track_position()
-            .custom_speed(1.0)
+            .custom_speed(1.0, SpecificType::soundtouch(options.soundtouch))
             .amplify(1.0)
             .pausable(false)
             .skippable()
@@ -163,7 +174,26 @@ impl Sink {
             });
 
         self.sound_count.fetch_add(1, Ordering::Relaxed);
-        let source = Done::new(source, self.sound_count.clone());
+
+        // the following allows us to avoid having to have a thread that waits for each sound to end to send a signal
+        // now with this, we can do much more stuff directly without having to have a extra thread for each sound.
+        let sound_count = self.sound_count.clone();
+        let pcmd_tx = self.pcmd_tx.clone();
+        let picmd_tx = self.picmd_tx.clone();
+        let source = source.cbdone(move || {
+            // the original function of `rodio::source::Done`, but we want to do more thant that
+            sound_count.fetch_sub(1, Ordering::Relaxed);
+
+            // using ".is_err()" here as the only error that can come from this channel is "Channel Closed"
+            if pcmd_tx.send(PlayerCmd::Eos).is_err() {
+                // not high priority, may log this on graceful exit because stop and player loop exit are not waiting on each-other
+                debug!("Player Channel is closed");
+            }
+            if picmd_tx.send(PlayerInternalCmd::Eos).is_err() {
+                error!("Player Internal Channel is closed");
+            }
+        });
+
         *self.sleep_until_end.lock() = Some(self.queue_tx.append_with_signal(source));
     }
 
@@ -304,32 +334,6 @@ impl Sink {
     #[inline]
     pub fn elapsed(&self) -> Duration {
         *self.controls.position.read()
-    }
-
-    // Spawns a new thread to sleep until the sound ends, and then sends the SoundEnded
-    // message through the given Sender.
-    pub fn message_on_end(&self) {
-        // TODO: we should not expose this function to be called outside of rusty-mod and always just have it active as we always need it
-        // NOTE: (in addition to the todo), we always use "message_on_end" to rely on EOS, so it is basically always set, so it should not be necessary
-        // to be set by commands every time, and worst, forget to call it or have race conditions that enqueue multiple things and wait on the wrong thing.
-        if let Some(sleep_until_end) = self.sleep_until_end.lock().take() {
-            let pcmd_tx = self.pcmd_tx.clone();
-            let picmd_tx = self.picmd_tx.clone();
-            std::thread::Builder::new()
-                .name("rusty message_on_end".into())
-                .spawn(move || {
-                    let _drop = sleep_until_end.recv();
-                    // using ".is_err()" here as the only error that can come from this channel is "Channel Closed"
-                    if pcmd_tx.send(PlayerCmd::Eos).is_err() {
-                        // not high priority, may log this on graceful exit because stop and player loop exit are not waiting on each-other
-                        debug!("Player Channel is closed");
-                    }
-                    if picmd_tx.send(PlayerInternalCmd::Eos).is_err() {
-                        error!("Player Internal Channel is closed");
-                    }
-                })
-                .expect("failed to spawn message_on_end thread");
-        }
     }
 }