+++ playback/src/backends/rusty/source/mod.rs
@@ -1,11 +1,13 @@
 //! Custom rodio sources and extension trait
 
+pub use custom_speed::SpecificType;
 use rodio::{Sample, Source};
 
 #[cfg(feature = "rusty-soundtouch")]
 pub mod soundtouch;
 
 pub mod async_ring;
+mod cb_done;
 mod custom_speed;
 
 /// Our sample type we choose to use across all places
@@ -19,12 +21,24 @@ where
     Self::Item: Sample,
 {
     /// A custom [`Source`] implementation to abstract away which speed module gets chosen.
-    fn custom_speed(self, initial_speed: f32) -> custom_speed::CustomSpeed<Self>
+    fn custom_speed(
+        self,
+        initial_speed: f32,
+        specific: SpecificType,
+    ) -> custom_speed::CustomSpeed<Self>
     where
         Self: Sized,
         Self: Source<Item = f32>,
     {
-        custom_speed::custom_speed(self, initial_speed)
+        custom_speed::custom_speed(self, initial_speed, specific)
+    }
+
+    /// Run a function once at the end of a source.
+    fn cbdone<Fn: FnOnce()>(self, fun: Fn) -> cb_done::CbDone<Self, Fn>
+    where
+        Self: Sized,
+    {
+        cb_done::CbDone::new(self, fun)
     }
 }