+++ playback/src/lib.rs
@@ -157,12 +157,12 @@ impl GeneralPlayer {
         stream_tx: StreamTX,
         playlist: SharedPlaylist,
     ) -> Result<Self> {
-        let config_read = config.read();
-        let backend = Backend::new_select(backend, &config_read, cmd_tx.clone());
+        let backend = Backend::new_select(backend, config.clone(), cmd_tx.clone());
 
         let db_path = get_app_config_path().with_context(|| "failed to get podcast db path.")?;
 
         let db_podcast = DBPod::new(&db_path).with_context(|| "error connecting to podcast db.")?;
+        let config_read = config.read();
         let db = DataBase::new(&config_read)?;
 
         let mpris = if config.read().settings.player.use_mediacontrols {
@@ -295,10 +295,6 @@ impl GeneralPlayer {
                 drop(playlist);
                 self.current_track_updated = true;
                 info!("gapless next track played");
-                #[allow(irrefutable_let_patterns)]
-                if let Backend::Rusty(ref mut backend) = self.backend {
-                    backend.message_on_end();
-                }
                 self.add_and_play_mpris_discord();
                 return;
             }
@@ -312,10 +308,6 @@ impl GeneralPlayer {
 
             self.add_and_play_mpris_discord();
             self.player_restore_last_position();
-            #[allow(irrefutable_let_patterns)]
-            if let Backend::Rusty(ref mut backend) = self.backend {
-                backend.message_on_end();
-            }
 
             self.send_stream_ev(UpdateEvents::TrackChanged(TrackChangedInfo {
                 current_track_index: u64::try_from(self.playlist.read().get_current_track_index())